requirements.txt
# Core dependencies
aiogram==3.7.0
aiohttp==3.9.5
python-dotenv==1.0.1

# Database
aiosqlite==0.20.0
sqlalchemy==2.0.30

# Image processing
Pillow==10.3.0

# Logging
loguru==0.7.2

# AI and ML dependencies
google-generativeai==0.5.4  # Для работы с Gemini AI
replicate==0.25.1          # Для генерации изображений через API
numpy==1.26.4              # Для A/B тестирования (Thompson Sampling)

# Translation
deep-translator==1.11.4    # Для автоматического перевода

# Speech-to-Text dependencies
openai-whisper==20231117   # Для распознавания речи
pydub==0.25.1             # Для обработки аудио файлов
ffmpeg-python==0.2.0      # Для конвертации аудио

# Development dependencies (optional)
# pytest==8.2.0
# pytest-asyncio==0.23.6
# black==24.4.2
# flake8==7.0.0

# Optional performance improvements
# uvloop==0.19.0  # Для Linux/Mac - ускоряет asyncio

# Старые ML dependencies (больше не нужны, используем Replicate API)
# diffusers==0.27.2
# transformers==4.40.2
# torch>=2.0.0
# torchvision
# accelerate==0.30.1

# Background removal (optional, когда понадобится)
# rembg==2.0.56